feat: add default persistent commands
Persistent commands are non-ephemeral (persistent) by default and need a silent option specified to be made ephemeral. All other commands are ephemeral by default. This allows certain commands to display a persistent message on their followup.
The silent option will still work regardless of whether the command is persistent or not.